se implementan sesiones de usuario, validaciones y cerrar sesion

+++ src/routes/contactos.js
@@ -16,15 +16,27 @@ const isAuth = (req, res, next) => {
 
 
 router.post('/insertar_contacto', isAuth, async (req, res) => {
-    const datosFormulario = req.body;
-    var insertar = {
-        nombre: datosFormulario.name
-        ,detalle: datosFormulario.detail
-        ,numero: datosFormulario.number
-    };
-    const contactos = new Contactos(insertar);
-    await contactos.save();
-    res.redirect('/listar_contactos');
+
+    let sesion = await Sistema.findOne({idSesion: req.session.id});
+
+    if(sesion){
+        let idUsuario = sesion.idUsuario;
+
+        const datosFormulario = req.body;
+        var insertar = {
+            idUsuario: idUsuario
+            ,nombre: datosFormulario.name
+            ,detalle: datosFormulario.detail
+            ,numero: datosFormulario.number
+        };
+        const contactos = new Contactos(insertar);
+        await contactos.save();
+        return res.redirect('/listar_contactos');
+        
+    }
+
+    return res.redirect('/');
+
 });
 
 router.get('/insertar_contacto_formulario', isAuth, (req, res) => {
@@ -33,35 +45,66 @@ router.get('/insertar_contacto_formulario', isAuth, (req, res) => {
 
 router.get('/listar_contactos', isAuth, async (req, res) => {
 
-    let sesion = Sistema.find()
+    let sesion = await Sistema.findOne({idSesion: req.session.id});
 
-    let contactos = await Contactos.find();
-    res.render('listar_contactos', {contactos});
+    if(sesion){
+
+        let contactos = await Contactos.find({idUsuario: sesion.idUsuario});
+
+        return res.render('listar_contactos', {contactos});
+    }
+
+
+    return res.redirect('/');
+
+   
 });
 
 router.post('/modificar_contacto_formulario/', isAuth, async (req, res) => {n
     //res.send(req.body.numero);
-    const datosContacto = req.body;
-    const contacto = await Contactos.findOne({_id: datosContacto.id});
-    res.render('modificar_contacto_formulario', {contacto});
+
+    let sesion = await Sistema.findOne({idSesion: req.session.id});
+
+    if(sesion){
+        const datosContacto = req.body;
+        const contacto = await Contactos.findOne({_id: datosContacto.id});
+        res.render('modificar_contacto_formulario', {contacto});
+    }
+    return res.redirect('/');
+
 });
 
 router.post('/modificar_contacto', isAuth, async (req, res) => {
-    const datosFormulario = req.body;
-    const datosEditados = {
-        nombre: datosFormulario.name,
-        detalle: datosFormulario.detail,
-        numero: datosFormulario.number
+    let sesion = await Sesion.findOne({idSesion: req.session.id});
+
+    if(sesion){
+
+        const datosFormulario = req.body;
+        const datosEditados = {
+            nombre: datosFormulario.name,
+            detalle: datosFormulario.detail,
+            numero: datosFormulario.number
+        }
+        await Contactos.updateOne({_id: datosFormulario.id}, datosEditados);
+        return res.redirect('/listar_contactos');
+
     }
-    await Contactos.updateOne({_id: datosFormulario.id}, datosEditados);
-    res.redirect('/listar_contactos');
+
+    return res.redirect('/');
+
 });
 
 router.post('/eliminar_contacto/', isAuth, async (req, res) => {
-    const { id }  = req.body;
-    let eliminar = await Contactos.deleteOne({_id: id});
-    //res.json({"redirect": "listar_contactos", "estado":eliminar});
-    res.redirect('/listar_contactos');
+    let sesion = await Sesion.findOne({idSesion: req.session.id});
+
+    if(sesion){
+        const { id }  = req.body;
+        let eliminar = await Contactos.deleteOne({_id: id});
+        //res.json({"redirect": "listar_contactos", "estado":eliminar});
+        return res.redirect('/listar_contactos');
+    }
+
+    return res.redirect('/');
 });
  
 module.exports = router;
